+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [#183](https://github.com/intridea/hashie/pull/183): Added Mash#load with YAML file support - [@gregory](https://github.com/gregory).
 * [#197](https://github.com/intridea/hashie/pull/197): Dont convert keys to string on initalization of mash - [@gregory](https://github.com/gregory).
+* [#198](https://github.com/intridea/hashie/pull/198): Add Smash and allow Mash to overwrite methods with `hash_<method>` aliasing - [@michaeherold](https://github.com/michaelherold).
 * Your contribution here.
 
 ## 3.2.0 (7/10/2014)